# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: All-Aboard API
  description: |
    Contrat HTTP Phase 2 MVP (feed, auth JWT, demandes d'aide, mentor feed).
    Source de vérité types : `packages/types`. Doc opérationnelle : `Docs/guides/web-api-integration.md`.
  version: 0.10.0
  contact:
    name: All-Aboard THP
    url: https://github.com/AllAboard-THP/All-Aboard
servers:
  - url: http://localhost:4000
    description: Développement local
  - url: https://api-staging.allaboard.fr
    description: Staging Dokploy
...
paths:
...
  /help-requests:
    post:
      operationId: postHelpRequests
      tags: [help-requests]
      summary: Créer une demande d'aide
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHelpRequestBody"
      responses:
        "201":
          description: Demande créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateHelpRequestResponse"
        "400":
          description: Corps invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Titre en doublon (normalisation espaces/casse)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuplicateErrorResponse"
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /help-requests/{id}:
    get:
      operationId: getHelpRequestById
      tags: [help-requests]
      summary: Détail d'une demande (public ; filtre certifications mentor optionnel)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filterByCertifications
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: >
            Si `true`, exige JWT mentor et ne renvoie que les réponses dont l'auteur
            partage au moins un tag avec la demande, ou est l'auteur de la demande.
      responses:
        "200":
          description: "Détail avec réponses (ordre chronologique)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequestDetailResponse"
        "401":
          description: JWT absent (filtre certifications)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Rôle non mentor (filtre certifications)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Demande absente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: not_found
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: patchHelpRequestById
      tags: [help-requests]
      summary: Modifier une demande (auteur JWT)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHelpRequestBody"
      responses:
        "200":
          description: Demande mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateHelpRequestResponse"
        "400":
          description: Corps invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Non auteur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Demande absente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteHelpRequestById
      tags: [help-requests]
      summary: Supprimer une demande (soft delete — auteur ou admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Demande supprimée (masquée du fil public)
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Non auteur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Demande absente ou déjà supprimée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /help-requests/{id}/help-mentor:
    post:
      operationId: postHelpRequestHelpMentor
      tags: [help-requests]
      summary: Demander l'aide d'un mentor (auteur JWT)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Flag mentor_help_requested activé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateHelpRequestResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Non auteur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Demande absente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /help-requests/{id}/responses:
    post:
      operationId: createHelpRequestResponse
      tags: [help-requests]
      summary: Ajouter une réponse (JWT requis)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResponseBody"
      responses:
        "201":
          description: Réponse créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateResponseResponse"
        "400":
          description: Corps invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: JWT absent ou invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Demande absente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: not_found
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /help-requests/{id}/responses/{responseId}:
    patch:
      operationId: patchHelpRequestResponse
      tags: [help-requests]
      summary: Modifier une réponse (auteur JWT)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: responseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResponseBody"
      responses:
        "200":
          description: Réponse mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResponseResponse"
        "400":
          description: Corps invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Non auteur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Réponse absente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteHelpRequestResponse
      tags: [help-requests]
      summary: Supprimer une réponse (auteur JWT)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: responseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Réponse supprimée
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Non auteur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Réponse absente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT émis par POST /auth/login (claim sub = userId)
...
  schemas:
...
    HelpRequestDetailResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/HelpRequest"
        responses:
          type: array
          items:
            $ref: "#/components/schemas/Response"
        certificationFilter:
          type: object
          required: [applied, totalCount, visibleCount]
          properties:
            applied:
              type: boolean
              const: true
            totalCount:
              type: integer
              minimum: 0
            visibleCount:
              type: integer
              minimum: 0
...
    CreateHelpRequestBody:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        tags:
          type: array
          maxItems: 32
          items:
            type: string
            maxLength: 64
        body:
          type: string
          maxLength: 50000
        codeSnippet:
          type: string
          maxLength: 50000
        codeLanguage:
          type: string
          maxLength: 64
        subjectId:
          type: string
          format: uuid
        urgent:
          type: boolean
        educationLevel:
          type: string
          maxLength: 64
    UpdateHelpRequestBody:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        body:
          type: string
          maxLength: 50000
        codeSnippet:
          type: string
          maxLength: 50000
          nullable: true
        codeLanguage:
          type: string
          maxLength: 64
        tags:
          type: array
          maxItems: 32
          items:
            type: string
            maxLength: 64
        subjectId:
          type: string
          format: uuid
          nullable: true
        urgent:
          type: boolean
        status:
          type: string
          enum: [open, resolved]
        educationLevel:
          type: string
          maxLength: 64
    UpdateHelpRequestResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/HelpRequest"
    CreateHelpRequestResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/HelpRequest"
        hints:
          type: object
          description: >
            Présent si l'agent All-Aboard (`POST /routing/evaluate`, via `AGENT_URL`)
            suggère un handoff Rubberduck (service externe).
          properties:
            rubberduckEligible:
              type: boolean
              const: true
    CreateResponseBody:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 10000
    CreateResponseResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/Response"
    UpdateResponseBody:
      type: object
      minProperties: 1
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 10000
        codeSnippet:
          type: string
          maxLength: 50000
          nullable: true
        codeLanguage:
          type: string
          maxLength: 64
    UpdateResponseResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/Response"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DuplicateErrorResponse:
      type: object
      required: [error, existingId]
      properties:
        error:
          type: string
          const: duplicate
        existingId:
          type: string
          format: uuid